Decouple compiler and environment
