+++ src/main/java/com/dukeacademy/solution/Compiler.java
@@ -1,12 +1,7 @@
 package com.dukeacademy.solution;
 
 import com.dukeacademy.solution.exceptions.CompilerFileContentException;
-import com.dukeacademy.commons.core.LogsCenter;
-import com.dukeacademy.solution.environment.CompilerEnvironment;
-import com.dukeacademy.solution.exceptions.CompilerEnvironmentException;
 import com.dukeacademy.solution.exceptions.CompilerException;
-import com.dukeacademy.solution.exceptions.CompilerFileCreationException;
-import com.dukeacademy.model.solution.UserProgram;
 import com.dukeacademy.solution.models.ClassFile;
 import com.dukeacademy.solution.models.JavaFile;
 
@@ -16,43 +11,20 @@ import javax.tools.JavaCompiler;
 import javax.tools.JavaFileObject;
 import javax.tools.StandardJavaFileManager;
 import javax.tools.ToolProvider;
-import java.io.File;
 import java.io.FileNotFoundException;
 import java.util.List;
 import java.util.Optional;
-import java.util.logging.Logger;
 
 public class Compiler {
     private static final String MESSAGE_COMPILER_FAILED = "Compiler failed.";
-    private static final String MESSAGE_FAILED_TO_CLEAR_ENVIRONMENT = "Failed to clear the compiler environment. "
-            + "Generated files and classes may persist.";
-
-    private static final Logger logger = LogsCenter.getLogger(Compiler.class);
-    private CompilerEnvironment environment;
     private JavaCompiler javaCompiler;
 
-    public Compiler(CompilerEnvironment environment) {
+    public Compiler() {
         this.javaCompiler = ToolProvider.getSystemJavaCompiler();
-        this.environment = environment;
     }
 
-    public void close() {
+    public ClassFile compileProgram(JavaFile javaFile) throws CompilerException, CompilerFileContentException {
         try {
-            this.environment.close();
-        } catch (CompilerEnvironmentException e) {
-            logger.fine(MESSAGE_FAILED_TO_CLEAR_ENVIRONMENT);
-        }
-    }
-
-    public ClassFile compileProgram(UserProgram program) throws CompilerException, CompilerFileContentException {
-        try {
-            environment.clearEnvironment();
-
-            String className = program.getClassName();
-            String code = program.getSourceCodeAsString();
-
-            JavaFile javaFile = environment.createJavaFile(className, code);
-
             DiagnosticCollector<JavaFileObject> diagnostics = new DiagnosticCollector<>();
             StandardJavaFileManager fileManager = javaCompiler.getStandardFileManager(diagnostics, null, null);
             Iterable<? extends JavaFileObject> sources = fileManager.getJavaFileObjects(javaFile.getFile());
@@ -65,30 +37,12 @@ public class Compiler {
             Optional<Diagnostic<? extends JavaFileObject>> error = errors.stream().findFirst();
             if (error.isPresent()) {
                 String errorMessage = error.map(diagnostic -> diagnostic.getMessage(null)).get();
-                this.clearEnvironmentAfterUserProgramInvalid(errorMessage);
                 throw new CompilerFileContentException(errorMessage);
             }
 
             return new ClassFile(javaFile.getCanonicalName(), javaFile.getClassPath());
-        } catch (CompilerFileCreationException | CompilerEnvironmentException |FileNotFoundException e) {
-            this.clearEnvironmentAfterCompilerFail();
+        } catch (FileNotFoundException e) {
             throw new CompilerException(MESSAGE_COMPILER_FAILED, e);
-        }
-    }
-
-    private void clearEnvironmentAfterUserProgramInvalid(String errorMessage) throws CompilerFileContentException {
-        try {
-            this.environment.clearEnvironment();
-        } catch (CompilerEnvironmentException e) {
-            throw new CompilerFileContentException(errorMessage + " " + MESSAGE_FAILED_TO_CLEAR_ENVIRONMENT);
-        }
-    }
-
-    private void clearEnvironmentAfterCompilerFail() throws CompilerException {
-        try {
-            this.environment.clearEnvironment();
-        } catch (CompilerEnvironmentException e) {
-            throw new CompilerException(MESSAGE_COMPILER_FAILED + " " + MESSAGE_FAILED_TO_CLEAR_ENVIRONMENT);
         }
     }
 }
