+++ src/test/java/com/dukeacademy/solution/CompilerTest.java
@@ -1,14 +1,19 @@
 package com.dukeacademy.solution;
 
-import com.dukeacademy.solution.environment.StandardCompilerEnvironment;
 import com.dukeacademy.solution.exceptions.CompilerException;
-import com.dukeacademy.solution.exceptions.CompilerEnvironmentException;
 import com.dukeacademy.solution.exceptions.CompilerFileContentException;
 import com.dukeacademy.model.solution.UserProgram;
+import com.dukeacademy.solution.exceptions.CompilerFileCreationException;
+import com.dukeacademy.solution.models.JavaFile;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.OutputStreamWriter;
+import java.io.Writer;
 import java.nio.file.Path;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -16,18 +21,11 @@ import static org.junit.jupiter.api.Assertions.*;
 class CompilerTest {
     @TempDir
     Path tempFolder;
-
     private Path environmentPath;
+
     private Compiler compiler;
 
     private UserProgram validProgram = new UserProgram("ValidTest", "public class ValidTest {\n"
-            + "\tpublic static void main(String args[]) {"
-            + "\t\tSystem.out.println(\"Hello world\");"
-            + "\t}"
-            + "\n}");
-    ;
-    private UserProgram validProgram1 = new UserProgram("ValidTest1", "package foo.bar;\n"
-            + "public class ValidTest1 {\n"
             + "\tpublic static void main(String args[]) {"
             + "\t\tSystem.out.println(\"Hello world\");"
             + "\t}"
@@ -36,28 +34,16 @@ class CompilerTest {
     private UserProgram invalidProgram = new UserProgram("InvalidTest", "FooBar");
 
     @BeforeEach
-    void initializeTest() throws CompilerEnvironmentException {
+    void initializeTest() {
+        compiler = new Compiler();
         environmentPath = tempFolder.resolve("compiler");
-        TestEnvironment environment = new TestEnvironment(environmentPath);
-        compiler = new Compiler(environment);
+        environmentPath.toFile().mkdir();
     }
 
     @Test
-    void compilerCreation() {
-        assertTrue(environmentPath.toFile().exists());
-    }
-
-    @Test
-    void close() throws CompilerException, CompilerFileContentException {
-        compiler.compileProgram(validProgram);
-        compiler.close();
-
-        assertFalse(environmentPath.toFile().exists());
-    }
-
-    @Test
-    void compileProgram() throws CompilerException, CompilerFileContentException {
-        compiler.compileProgram(validProgram);
+    void compileProgram() throws CompilerException, CompilerFileContentException, IOException, CompilerFileCreationException {
+        JavaFile validJavaFile = this.createJavaFile(validProgram);
+        compiler.compileProgram(validJavaFile);
 
         Path validJavaFilePath = environmentPath.resolve(validProgram.getClassName() + ".java");
         Path validClassFilePath = environmentPath.resolve(validProgram.getClassName() + ".class");
@@ -65,21 +51,31 @@ class CompilerTest {
         assertTrue(validJavaFilePath.toFile().exists());
         assertTrue(validClassFilePath.toFile().exists());
 
-        compiler.compileProgram(validProgram1);
-        Path validJavaFilePath1 = environmentPath.resolve("foo").resolve("bar").resolve("ValidTest1" + ".java");
-        Path validClassFilePath1 = environmentPath.resolve("foo").resolve("bar").resolve("ValidTest1" + ".class");
-
-        assertTrue(validJavaFilePath1.toFile().exists());
-        assertTrue(validClassFilePath1.toFile().exists());
-
-        assertThrows(CompilerFileContentException.class, () -> compiler.compileProgram(invalidProgram));
-
-        Path invalidJavaFilePath = environmentPath.resolve(invalidProgram.getClassName() + ".java");
+        JavaFile invalidJavaFile = this.createJavaFile(invalidProgram);
+        assertThrows(CompilerFileContentException.class, () -> compiler.compileProgram(invalidJavaFile));
         Path invalidClassFilePath = environmentPath.resolve(invalidProgram.getClassName() + ".class");
 
-        assertFalse(invalidJavaFilePath.toFile().exists());
         assertFalse(invalidClassFilePath.toFile().exists());
 
-        compiler.close();
+    }
+
+    private JavaFile createJavaFile(UserProgram program) throws IOException, CompilerFileCreationException {
+        String path = environmentPath.resolve(program.getClassName() + ".java").toUri().getPath();
+        File file = new File(path);
+
+        boolean isFileCreated = file.createNewFile();
+
+        if (!isFileCreated) {
+            throw new CompilerFileCreationException("Unable to create file");
+        }
+
+        FileOutputStream fileOutputStream = new FileOutputStream(file);
+        Writer fileWriter = new OutputStreamWriter(fileOutputStream);
+
+        fileWriter.write(program.getSourceCodeAsString());
+
+        fileWriter.close();
+
+        return new JavaFile(program.getClassName(), environmentPath.toUri().getPath());
     }
 }
